Moved LoginSignstyle and updated imports
