+++ src/pages/Login/LoginSignStyle.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
-import { theme } from './colors'
+import { theme } from '../../utils/style/colors'
 import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
 import VisibilityIcon from '@mui/icons-material/Visibility'
